Refactored UI code. Removed redundant logic. Updated comments.
