+++ frontend/app/theme/styleConstants.ts
@@ -1,4 +1,6 @@
-// sharedStyles.ts – Common style objects to be reused across components
+// styleConstants.ts – Consolidated UI constants and shared style objects used across components.
+
+export const TRANSITION_SPEED_FAST = "0.10s";
 
 export const commonTransition = '0.2s ease';
 export const commonPadding = '8px';
