--- frontend/app/theme/uiConstants.ts
@@ -1,3 +1,0 @@
-// uiConstants.ts – Defines application-wide UI constants like animation speeds and timing values.
-
-export const TRANSITION_SPEED_FAST = "0.10s";
